smart-city-iot/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
paho-mqtt==2.1.0
SQLAlchemy>=2.0.35
python-dotenv==1.0.1